python/requirements.txt
pytest==7.4.3
pytest-json==0.4.0
requests==2.31.0
pytest-html==4.1.1
mysql-connector-python==8.2.0